refactor: 大きめモジュールをディレクトリ構成へ分割して見通しを改善
windows-services ブランチ (#61) を土台に、肥大化したフラットなモジュール
ファイルを責務ごとのサブモジュールへ分割した。公開 API（`crate::config::*`
や `crate::logger::*` 等）は mod.rs での再エクスポートで完全に維持しており、
挙動は不変。テスト 77 件はすべて緑のまま。

- config.rs (669行) → config/{types,model,loader,validate,tests}.rs
  列挙型・データ構造・読み込み・バリデーションを分離。
- logger.rs (606行) → logger/{format,writer,tests}.rs ＋ mod.rs
  整形/レベル判定（format）と書き込みタスク（writer）を分離。
- dashboard.rs (652行) → dashboard/{event,server,search,tests}.rs ＋ mod.rs
  配信イベント・HTTP/SSE サーバ・過去ログ検索を分離。dashboard.html も同梱移動。
- Windows 専用コード service.rs / win_runas.rs を platform/ 配下へ集約。
  参照を crate::platform::* へ更新（main / actions::spawn）。

あわせて軽微なクリーンアップ:
- writer の try_recv ループを while let 化、append_to_file を &Path 受けに変更。
- static_root_of_destination の冗長な let 束縛を除去。
- build.rs に CAT_WATCHER_SKIP_WINRES=1 を追加。Windows 以外のホストから
  `cargo check --target *-windows-*` で Windows 専用コードを型検査できるようにする
  （既定 off・本番ビルドのリソース埋め込みは不変）。
