+++ cat-watcher/src/logger/tests.rs
